requirements.txt
streamlit
agno
google-generativeai
google-search-results
yagmail
python-dotenv
